requirements.txt
google-generativeai==0.7.2
google-cloud-texttospeech==2.16.3
google-auth==2.33.0
chromadb==0.4.22
langchain==0.1.6
langchain-community==0.0.20
langchain-core==0.1.23
sentence-transformers==3.0.1
beautifulsoup4==4.12.3
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
python-dotenv==1.0.1
streamlit==1.37.1
html2text==2020.1.16
tiktoken==0.5.2
pygame==2.6.0